Add support for hybrid data type and SVD latent fusion in models
This update enhances the data handling capabilities in the image_data.py and base_gen2.py files, adding support for a hybrid datatype and improving the randomization function. A new HybridDataset class has been introduced, allowing for an additional input, p, in the randomize_inputs function. The introduced  SVD latent fusion in the EPRAUNe model allows for data processing efficiency. Moreover, an option has been added in the SVDDataset for creating fixed subsets of training and validation data.